pyproject.toml
[project]
name = "yandex-direct-metrica-mcp"
version = "0.1.0"
description = "MCP server for Yandex Direct + Metrica (read-only, raw exports)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.3.0",
    "tapi-yandex-direct",
    "tapi-yandex-metrika",
    "requests>=2.31.0",
    "orjson>=3.9.0",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
]

[project.scripts]
mcp-yandex-ad = "mcp_yandex_ad:main"
yandex-direct-metrica-mcp = "mcp_yandex_ad:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_yandex_ad"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mcp[cli]>=1.3.0",
]
